refactor: flatten folder structure
