+++ src/angular-public/usermenu/public-api.ts
@@ -1,10 +1,10 @@
 export * from './usermenu.module';
 // TODO: Deprecate UserMenuModule module usage (Only UsermenuModule should be supported)
 export { SbbUsermenuModule as UserMenuModule } from './usermenu.module';
-export * from './usermenu/usermenu.component';
+export * from './usermenu';
 /** @deprecated Remove with v12 */
 export { SbbUsermenuModule as UsermenuModule } from './usermenu.module';
 /** @deprecated Remove with v12 */
-export { SbbUserMenu as UserMenuComponent } from './usermenu/usermenu.component';
+export { SbbUserMenu as UserMenuComponent } from './usermenu';
 export * from './usermenu-animations';
-export * from './usermenu-item/usermenu-item';
+export * from './usermenu-item';
