+++ src/angular-public/usermenu/usermenu.scss
@@ -1,4 +1,4 @@
-@import '../../../angular-core/styles/common';
+@import '../../angular-core/styles/common';
 $textColor: if($sbbBusiness, $sbbColorIron, $sbbColorGrey);
 
 $triggerHeight: 48px;
